refactor: move column generation logic to composable
