+++ components/geojson-table-window-content.vue
@@ -1,13 +1,5 @@
 <script lang="ts" setup>
-import {
-	type AccessorColumnDef,
-	type CellContext,
-	type ColumnDef,
-	createColumnHelper,
-	type GroupColumnDef,
-	type Row,
-	type Table,
-} from "@tanstack/vue-table";
+import type { ColumnDef, Row, Table } from "@tanstack/vue-table";
 
 import { useGeojsonStore } from "@/stores/use-geojson-store.ts";
 import type { FeatureType, MarkerType } from "@/types/global";
@@ -19,83 +11,7 @@ const url = "https://raw.githubusercontent.com/wibarab/wibarab-data/main/wibarab
 const { isPending } = GeojsonStore.fetchGeojson(url);
 const { fetchedData, tables } = storeToRefs(GeojsonStore);
 const { data: projectData } = useProjectInfo();
-const columnHelper = createColumnHelper<FeatureType>();
-const { AND_OPERATOR } = useAdvancedQueries();
-
-interface SimpleColumnInterface {
-	id: string;
-	header: string;
-	columns: Array<AccessorColumnDef<FeatureType> | GroupColumnDef<FeatureType>>;
-	enableHiding?: boolean;
-}
-
-function buildColumnDefRecursive(
-	col: SimpleColumnInterface,
-	featureCategories: Record<string, string>,
-	allFeatureNames: Array<Record<string, string>>,
-): SimpleColumnInterface {
-	const subCategories = Object.entries(featureCategories).filter(
-		([categoryName, _]) =>
-			categoryName.startsWith(col.id) && categoryName.lastIndexOf(".") === col.id.length,
-	);
-	let columns: Array<AccessorColumnDef<FeatureType> | GroupColumnDef<FeatureType>> = [];
-	columns = columns.concat(
-		allFeatureNames
-			.filter((heading) => heading.category === col.id)
-			.map((heading) => {
-				const accessorFn = (cell: FeatureType) => {
-					return Object.keys(
-						cell.properties[String(Object.keys(heading).find((key) => /ft_*/.test(key)) ?? "")] ??
-							{},
-					);
-				};
-				return columnHelper.accessor(accessorFn, {
-					id: Object.keys(heading).find((key) => /ft_*/.test(key)) ?? "",
-					header: heading[Object.keys(heading).find((key) => /ft_*/.test(key)) ?? ""],
-					cell: (cell: CellContext<FeatureType, unknown>) => {
-						return h(resolveComponent("GeojsonTablePropertyCell"), {
-							value: cell.row.original.properties[cell.column.columnDef.id!],
-						});
-					},
-					filterFn: (row, columnId, filterValue: Map<string, unknown>) => {
-						if (!row.getVisibleCells().find((cell) => cell.column.id === columnId)) {
-							return true;
-						}
-						if (filterValue.size === 0) return true;
-						const filter = [...filterValue.keys()].some((val) => {
-							if (val.includes(AND_OPERATOR)) {
-								return val
-									.split(AND_OPERATOR)
-									.every((v) => (row.getValue(columnId) as Array<string>).includes(v));
-							} else return (row.getValue(columnId) as Array<string>).includes(val);
-						});
-						return filter;
-					},
-					enableGlobalFilter: true,
-				}) as AccessorColumnDef<FeatureType>;
-			}),
-	);
-
-	columns = columns.concat(
-		subCategories
-			.map(([categoryName, categoryLabel]) => {
-				return columnHelper.group(
-					buildColumnDefRecursive(
-						{
-							header: featureCategories[categoryName] ?? categoryLabel,
-							id: String(categoryName),
-							columns: [],
-						},
-						featureCategories,
-						allFeatureNames,
-					),
-				) as GroupColumnDef<FeatureType>;
-			})
-			.filter((col) => (col.columns?.length ?? 0) > 0),
-	);
-
-	return { ...col, columns: columns };
-}
+const { createColumnDefs } = useColumnGeneration();
 
 const columns = computed(() => {
 	const allFeatureNames = fetchedData.value.get(url)?.properties.column_headings;
@@ -105,70 +21,9 @@ const columns = computed(() => {
 		string
 	>;
 
-	const topLevelColumns: Array<SimpleColumnInterface> = [
-		{
-			id: "-",
-			header: "-",
-			enableHiding: false,
-			columns: [],
-		},
-		...[
-			...new Set(Object.keys(featureCategories).map((categoryName) => categoryName.split(".")[0]!)),
-		].map((categoryName) => {
-			return {
-				header: featureCategories[categoryName] ?? categoryName,
-				id: categoryName,
-				columns: [],
-			};
-		}),
-	];
-
-	topLevelColumns.forEach((col) => {
-		let subcategoryColumns;
-		if (col.header !== "-") {
-			subcategoryColumns = buildColumnDefRecursive(
-				col,
-				featureCategories,
-				allFeatureNames!,
-			).columns;
-		} else {
-			subcategoryColumns = [
-				columnHelper.group({
-					header: "-",
-					id: "-",
-					enableHiding: false,
-					//@ts-expect-error type mismatch in accessorFn
-					columns: allFeatureNames
-						.filter((heading) => !heading.category)
-						.map((heading) => {
-							return {
-								id: Object.keys(heading)[0],
-								header: Object.values(heading)[0],
-								enableHiding: false,
-								cell: ({ cell }: CellContext<FeatureType, never>) => {
-									return h(
-										"span",
-										{ class: "max-w-[500px] truncate font-medium" },
-										cell.row.original.properties[cell.column.columnDef.id!],
-									);
-								},
-								accessorFn: (cell: FeatureType) => {
-									return cell.properties[String(Object.keys(heading)[0])];
-								},
-								enableColumnFilter: false,
-								enableGlobalFilter: true,
-							};
-						}),
-				}),
-			];
-		}
-		col.columns = subcategoryColumns;
-	});
-	const groupedColumns = topLevelColumns
-		// .filter((col) => col.columns.some((col) => (col.columns?.length ?? -1) > 0))
-		.map((col) => columnHelper.group(col))
-		.sort((a, b) => String(a.header).localeCompare(String(b.header)));
-	return groupedColumns;
+	if (allFeatureNames && featureCategories)
+		return createColumnDefs(featureCategories, allFeatureNames);
+	else return [];
 });
 const columnVisibility = computed(() => {
 	const columnHeadings = fetchedData.value.get(url)?.properties.column_headings ?? [];
